PC-Automation-main/requirements.txt
pyautogui
requests
beautifulsoup4
googlesearch-python
wmi
psutil
pyaudio
numpy
pywhatkit
scipy
comtypes
pycaw
fuzzywuzzy
